Moved table to its own component, used flexbox to make the table overflow-y:scroll properly.
Not happy about the lack of real estate the table now has, I liked having the entire screen for it. But alas.
